requirements.txt
apturl==0.5.2
bcrypt==3.2.0
bidict==0.23.1
blinker==1.7.0
Brlapi==0.8.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
Flask==3.0.2
flask-sock==0.7.0
Flask-SocketIO==5.3.6
fonttools==4.48.1
future==0.18.2
gyp==0.1
h11==0.14.0
httplib2==0.20.2
idna==3.6
importlib-metadata==7.0.1
iso8601==2.1.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
matplotlib==3.8.2
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
paramiko==2.9.3
pexpect==4.8.0
pillow==10.2.0
plotly==5.18.0
protobuf==3.12.4
ptyprocess==0.7.0
pyarrow==15.0.0
pycairo==1.20.1
pycups==2.0.1
pygame==2.5.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-engineio==4.9.0
python-socketio==5.11.2
PythonTurtle==0.3.2
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
retrying==1.3.4
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
simple-websocket==1.0.0
six==1.16.0
systemd-python==234
tenacity==8.2.3
tk==0.1.0
Turbo-Flask==0.8.4
typing_extensions==4.9.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.0
usb-creator==0.3.7
wadllib==1.3.6
Werkzeug==3.0.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==3.17.0
